Refactoring evaluation
